merge

+++ src/main/java/com/tplus/gwland/svy/repository/SurveyRepository.java
@@ -1,3 +1,4 @@
+
 package com.tplus.gwland.svy.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -6,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import com.tplus.gwland.svy.domain.Survey;
 
 @Repository
-public interface SurveyRepository extends JpaRepository<Survey, Long>{
+public interface SurveyRepository extends JpaRepository<Survey, Integer>{
 
 }
